LIPAMPIRI/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
python-dotenv==1.0.0
phonenumbers 
django-phonenumber-field
django-countries
python-dateutil
# mysqlclient==2.2.4